* Remove: deprecated classes  * Refactor: renamed `UserID` into `UserId` to increase consitensy  * Tests: updated tests
