+++ src/main/scala/com/igeolise/slack/ApiSlackClient.scala
@@ -6,5 +6,6 @@ import com.igeolise.slack.dto.InteractiveMessage
 import scala.concurrent.Future
 
 trait ApiSlackClient {
+  val postMessageUrl = "https://slack.com/api/chat.postMessage"
   def sendInteractiveMessage(interactiveMessage: InteractiveMessage, authToken: String): Future[Either[Error, Unit]]
 }
