+++ src/main/scala/com/igeolise/slack/HooksSlackClient.scala
@@ -37,7 +37,7 @@ object HooksSlackClient {
   sealed abstract class Notify(val formatted: String)
   object Notify {
     case object Channel extends Notify("<!channel>")
-    case class UserID(id: String) extends Notify(s"<@$id>")
+    case class UserId(id: String) extends Notify(s"<@$id>")
     case class UserGroup(groupName: String, groupId: String) extends Notify(s"<!subteam^$groupId|$groupName>")
   }
 }
