+++ src/main/scala/com/igeolise/slack/util/PayloadMapper.scala
@@ -1,0 +1,24 @@
+package com.igeolise.slack.util
+
+import com.igeolise.slack.HooksSlackClient.HookMessage
+import play.api.libs.json.{JsObject, Json}
+
+object PayloadMapper {
+
+  def toBodyJson(hookMessage: HookMessage): JsObject = {
+    val notifyMapped =
+      hookMessage.notifications.map(_.formatted).mkString(" ") + { if(hookMessage.notifications.nonEmpty) " " else "" }
+
+    Json.obj(
+      "text" -> (notifyMapped + hookMessage.msg),
+      "attachments" -> hookMessage.attachments.map { attachment =>
+        Json.obj(
+          "text" -> attachment.text,
+          "color" -> attachment.color.code,
+          "mrkdwn_in" -> Seq("text")
+        )
+      },
+      "link_names" -> 1
+    )
+  }
+}
